search+search-api: Move aggregation building to `search`
Lets move the aggregation building to a shared library since we will be
doing the same thing in concept-api.
