+++ search-api/src/main/scala/no/ndla/searchapi/model/api/MultiSearchResult.scala
@@ -7,6 +7,7 @@
 
 package no.ndla.searchapi.model.api
 
+import no.ndla.search.api.MultiSearchTermsAggregation
 import org.scalatra.swagger.annotations.{ApiModel, ApiModelProperty}
 
 import scala.annotation.meta.field
