+++ search-api/src/main/scala/no/ndla/searchapi/service/search/MultiDraftSearchService.scala
@@ -17,6 +17,7 @@ import no.ndla.common.model.domain.draft.DraftStatus
 import no.ndla.language.Language.AllLanguages
 import no.ndla.language.model.Iso639
 import no.ndla.network.model.RequestInfo
+import no.ndla.search.AggregationBuilder.{buildTermsAggregation, getAggregationsFromResult}
 import no.ndla.search.Elastic4sClient
 import no.ndla.searchapi.Props
 import no.ndla.searchapi.model.api.ErrorHelpers
@@ -108,7 +109,7 @@ trait MultiDraftSearchService {
         Failure(ResultWindowTooLargeException())
       } else {
 
-        val aggregations = buildTermsAggregation(settings.aggregatePaths)
+        val aggregations = buildTermsAggregation(settings.aggregatePaths, indexServices.map(_.getMapping))
 
         val searchToExecute = search(searchIndex)
           .query(filteredSearch)
