+++ search/src/main/scala/no/ndla/search/model/domain/TermAggregation.scala
@@ -1,11 +1,11 @@
 /*
- * Part of NDLA search-api
- * Copyright (C) 2021 NDLA
+ * Part of NDLA search
+ * Copyright (C) 2024 NDLA
  *
  * See LICENSE
  */
 
-package no.ndla.searchapi.model.domain
+package no.ndla.search.model.domain
 
 case class Bucket(value: String, count: Int)
 case class TermAggregation(
